Remove the need for after evaluate block to choose between jar and war publications
