+++ src/main/groovy/nebula/plugin/publishing/ivy/IvyBasePublishPlugin.groovy
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 package nebula.plugin.publishing.ivy
-
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.XmlProvider
@@ -23,13 +22,32 @@ import org.gradle.api.plugins.WarPlugin
 import org.gradle.api.publish.ivy.IvyPublication
 
 class IvyBasePublishPlugin implements Plugin<Project> {
+    static final String IVY_WAR = 'nebulaPublish.ivy.war'
+    static final String IVY_JAR = 'nebulaPublish.ivy.jar'
+
     @Override
     void apply(Project project) {
         project.plugins.apply org.gradle.api.publish.ivy.plugins.IvyPublishPlugin
 
+        project.ext.set(IVY_WAR, false)
+        project.ext.set(IVY_JAR, false)
+
+        project.plugins.withType(WarPlugin) {
+            project.ext.set(IVY_WAR, true)
+        }
+
+        project.plugins.withType(JavaPlugin) {
+            project.ext.set(IVY_JAR, true)
+        }
+
         project.publishing {
             publications {
                 nebulaIvy(IvyPublication) {
+                    if (project.ext.get(IVY_WAR)) {
+                        from project.components.web
+                    } else if (project.ext.get(IVY_JAR)) {
+                        from project.components.java
+                    }
                     descriptor.status = project.status
                     descriptor.withXml { XmlProvider xml ->
                         def root = xml.asNode()
@@ -40,20 +58,6 @@ class IvyBasePublishPlugin implements Plugin<Project> {
                             infoNode = infoNode[0]
                         }
                         infoNode.appendNode('description', [:], project.description ?: '')
-                    }
-                }
-            }
-        }
-
-        project.afterEvaluate {
-            project.publishing {
-                publications {
-                    nebulaIvy(IvyPublication) {
-                        if (project.plugins.hasPlugin(WarPlugin)) {
-                            from project.components.web
-                        } else if (project.plugins.hasPlugin(JavaPlugin)){
-                            from project.components.java
-                        }
                     }
                 }
             }
