--- src/main/groovy/nebula/plugin/publishing/maven/MavenDependenciesJarPlugin.groovy
@@ -1,35 +1,0 @@
-/*
- * Copyright 2015 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package nebula.plugin.publishing.maven
-
-import org.gradle.api.Project
-import org.gradle.api.plugins.JavaPlugin
-import org.gradle.api.publish.maven.MavenPublication
-
-class MavenDependenciesJarPlugin extends MavenDependenciesBase {
-    @Override
-    void setupPublishingForProject(Project project) {
-        project.plugins.withType(JavaPlugin) {
-            project.publishing {
-                publications {
-                    nebula(MavenPublication) {
-                        from project.components.java
-                    }
-                }
-            }
-        }
-    }
-}
