--- src/main/resources/META-INF/gradle-plugins/nebula.maven-dependencies-jar.properties
@@ -1,1 +1,0 @@
-implementation-class=nebula.plugin.publishing.maven.MavenDependenciesJarPlugin
