--- src/main/resources/META-INF/gradle-plugins/nebula.maven-resolved-dependencies-jar.properties
@@ -1,1 +1,0 @@
-implementation-class=nebula.plugin.publishing.maven.MavenResolvedDependenciesJarPlugin
