--- src/test/groovy/nebula/plugin/publishing/maven/MavenDependenciesJarPluginIntegrationSpec.groovy
@@ -1,68 +1,0 @@
-/*
- * Copyright 2015 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package nebula.plugin.publishing.maven
-
-import nebula.test.IntegrationSpec
-
-
-class MavenDependenciesJarPluginIntegrationSpec extends IntegrationSpec {
-    File publishDir
-
-    def setup() {
-        buildFile << """\
-            ${applyPlugin(MavenDependenciesJarPlugin)}
-
-            version = '0.1.0'
-            group = 'test.nebula'
-
-            publishing {
-                repositories {
-                    maven {
-                        name = 'testLocal'
-                        url = 'testrepo'
-                    }
-                }
-            }
-        """.stripIndent()
-
-        settingsFile << '''\
-            rootProject.name = 'maventest'
-        '''.stripIndent()
-
-        publishDir = new File(projectDir, 'testrepo/test/nebula/maventest/0.1.0')
-    }
-
-    def 'creates a jar publication'() {
-        buildFile << '''\
-            apply plugin: 'java'
-        '''.stripIndent()
-
-        when:
-        runTasksSuccessfully('publishNebulaPublicationToTestLocalRepository')
-
-        then:
-        new File(publishDir, 'maventest-0.1.0.jar').exists()
-        new File(publishDir, 'maventest-0.1.0.pom').exists()
-    }
-
-    def 'does not fail on no java plugin'() {
-        when:
-        runTasksSuccessfully('publishNebulaPublicationToTestLocalRepository')
-
-        then:
-        noExceptionThrown()
-    }
-}
