--- src/test/groovy/nebula/plugin/publishing/maven/MavenResolvedDependenciesJarPluginIntegrationSpec.groovy
@@ -1,72 +1,0 @@
-/*
- * Copyright 2015 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package nebula.plugin.publishing.maven
-
-import nebula.test.IntegrationSpec
-import nebula.test.dependencies.DependencyGraphBuilder
-import nebula.test.dependencies.GradleDependencyGenerator
-
-class MavenResolvedDependenciesJarPluginIntegrationSpec extends IntegrationSpec {
-    File publishDir
-
-    def setup() {
-        buildFile << """\
-            ${applyPlugin(MavenResolvedDependenciesJarPlugin)}
-
-            version = '0.1.0'
-            group = 'test.nebula'
-
-            publishing {
-                repositories {
-                    maven {
-                        name = 'testLocal'
-                        url = 'testrepo'
-                    }
-                }
-            }
-        """.stripIndent()
-
-        settingsFile << '''\
-            rootProject.name = 'resolvedmaventest'
-        '''.stripIndent()
-
-        publishDir = new File(projectDir, 'testrepo/test/nebula/resolvedmaventest/0.1.0')
-    }
-
-    def 'dynamic versions are replaced by the resolved version'() {
-        def graph = new DependencyGraphBuilder().addModule('test.resolved:a:1.0.0')
-                .addModule('test.resolved:a:1.1.0').build()
-        File mavenrepo = new GradleDependencyGenerator(graph, "${projectDir}/testrepogen").generateTestMavenRepo()
-
-        buildFile << """\
-            apply plugin: 'java'
-
-            repositories { maven { url '${mavenrepo.absolutePath}' } }
-
-            dependencies {
-                compile 'test.resolved:a:1.+'
-            }
-        """.stripIndent()
-
-        when:
-        runTasksSuccessfully('publishNebulaPublicationToTestLocalRepository')
-
-        then:
-        def root = new XmlSlurper().parseText(new File(publishDir, 'resolvedmaventest-0.1.0.pom').text)
-        def dependency = root.dependencies.dependency[0]
-        dependency.version.text() == '1.1.0'
-    }
-}
